javascript: make deprecation notice more prominent
This text is easy to miss as it's only at the top, and in the same size
and weight as other intro text.  If people follow links to sections (as
is the norm for style guides), this will be completely skipped.

Let's make the notice into a sticky butter bar at the top so it's always
shown regardless of where the person is on the page.