jsguide: import last internal version
Internally, we've turned down the JS styleguide in favor of a merged
TS/JS guide (with heavy emphasis on TS).  That variant doesn't exist
publicly yet, and even then, we probably want to keep around the JS
version as-is more or less indefinitely like we have with the older
ES5 guide.  Once we update the TS guide, we can add a deprecation
notice to this file like we do with the ES5 guide.

There is one change that has been left out: the heavy emphasis on
only using clang-format as the rules are heavily geared towards
google3, and there is no external guidance for formatting.  If we
could provide public guidance here, it'd be great to update, but
since the guide is no more internally, that seems unlikely, so it's
probably best to freeze the pre-clang-format guidance here.